feat: Add List View screen and update Home Screen
This commit adds a new screen, List View, to the app. It includes importing the `ListViewScreen` class in `home_screen.dart` and adding a new `ListTile` to the `HomeScreen` widget for navigating to the List View screen. The route name for `ListViewScreen` is defined in `list_view_screen.dart`, and the `ListViewScreen` route is added to the `routes` map in `routes.dart`. These changes enhance the app's functionality by providing a new screen for displaying a list view.