feat: Add button to show Cupertino alert
This commit adds a button to the body of the `AlertsScreen` widget in `alerts_screen.dart`. The button is a `TextButton` widget with the text "Show cupertino alert". When pressed, the button will trigger an action. This change enhances the functionality of the app by providing a way to display a Cupertino alert.